fix: widen appliance uplink usage Sent/Received to long
Cumulative byte counts for a 31-day window exceed Int32.MaxValue on busy
uplinks, so the response failed to deserialize and the call threw.

Closes #360

+++ Meraki.Api/Data/OrganizationApplianceUplinksUsageByNetworkItemByUplinkItem.cs
@@ -11,14 +11,14 @@ public class OrganizationApplianceUplinksUsageByNetworkItemByUplinkItem
 	/// </summary>
 	[ApiAccess(ApiAccess.Read)]
 	[DataMember(Name = "received")]
-	public int Received { get; set; }
+	public long Received { get; set; }
 
 	/// <summary>
 	/// Bytes sent
 	/// </summary>
 	[ApiAccess(ApiAccess.Read)]
 	[DataMember(Name = "sent")]
-	public int Sent { get; set; }
+	public long Sent { get; set; }
 
 	/// <summary>
 	/// Uplink name
